requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
langchain==0.0.335
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
kaggle==1.5.16
pandas==2.1.3
numpy==1.26.2 